Make reviewed changes
