--- src/App.ts
@@ -1,3 +1,0 @@
-import pullIssuesAndCreateDatabase from "./PullIssuesAndCreateDatabase.ts";
-
-pullIssuesAndCreateDatabase(process.argv[2] , process.argv[3]);
