+++ src/big-query-functions/table-exists.ts
@@ -1,6 +1,6 @@
+import { BigQuery } from "@google-cloud/bigquery";
 //returns a promise with true if table exists, or false if not
 export default async function tableExists(datasetId: string, tableId: string) : Promise<boolean> {
-    const {BigQuery} = require('@google-cloud/bigquery');
     const bigquery = new BigQuery();
     const dataset = bigquery.dataset(datasetId);
